requirements.txt
python-dotenv
pandas 
numpy
pymongo
pymongo[srv]==3.6
certifi

-e .
